# public/openapi-everything.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AB Brain - Complete API",
...
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "https://ab-brain.up.railway.app/api"
    }
  ],
  "paths": {
...
    "/exercises": {
      "get": {
        "operationId": "listExercises",
        "summary": "Search the exercise catalog \u2014 1060+ exercises with equipment, muscle groups, level, and muscle_strength_score (mscore). Use mscore to pick the most effective exercises for a given muscle group.",
        "description": "Returns exercises sorted by muscle_strength_score (mscore) descending by default. Each exercise includes: name, level (beginner/intermediate/advanced), equipment needed, primary_muscle_groups, category, muscle_strength_score (0-100, higher=more effective), sets_logged (popularity), and description.\n\n**mscore ranges:** 90-100 = elite compound movements (primary lifts), 70-89 = strong accessories, 50-69 = moderate/isolation, below 50 = light/stabilizer.\n\n**sets_logged** = popularity across all Fitbod users. Higher = more commonly performed and validated.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Full-text search across name, equipment, muscle groups, description"
          },
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "beginner",
                "intermediate",
                "advanced"
              ]
            },
            "description": "Filter by difficulty level"
          },
          {
            "name": "equipment",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by equipment type (e.g. Barbell, Dumbbell, Cable, Body Weight, Kettlebell)"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by muscle category (e.g. chest, back, shoulders, legs, arms, core, lower_back)"
          },
          {
            "name": "muscle_group",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by primary muscle group name (partial match, e.g. 'chest', 'quadriceps', 'hamstrings')"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "mscore_desc",
                "mscore_asc",
                "name",
                "sets_logged",
                "level"
              ]
            },
            "description": "Sort order (default: mscore descending)"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise catalog results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "exercises": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Exercise"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/exercises/{id}": {
      "get": {
        "operationId": "getExercise",
        "summary": "Get a single exercise by ID with full details including description and mscore",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              }
            }
          }
        }
      }
    },
    "/gym-profiles": {
      "get": {
        "operationId": "listGymProfiles",
        "summary": "List all gym profiles with their equipment",
        "responses": {
          "200": {
            "description": "List of gym profiles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "gym_profiles": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GymProfile"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createGymProfile",
        "summary": "Create a gym profile with available equipment list",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GymProfileInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GymProfile"
                }
              }
            }
          }
        }
      }
    },
...
    "/gym-profiles/{id}": {
      "put": {
        "operationId": "updateGymProfile",
        "summary": "Update gym profile name, equipment, or primary status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GymProfileInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GymProfile"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteGymProfile",
        "summary": "Delete a gym profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Exercise": {
        "type": "object",
        "description": "An exercise from the Fitbod catalog. muscle_strength_score (mscore) is a 0-100 score of how effectively this exercise builds strength in its target muscle group. sets_logged is total sets logged across all Fitbod users (popularity indicator).",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Exercise name (e.g. 'Barbell Bench Press', 'Pull Up', 'Romanian Deadlift')"
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "description": "Difficulty level"
          },
          "equipment": {
            "type": "string",
            "description": "Equipment required (e.g. 'Barbell', 'Dumbbell', 'Cable', 'Body Weight', 'Kettlebell', 'Machine')"
          },
          "primary_muscle_groups": {
            "type": "string",
            "description": "Primary muscles targeted (e.g. 'chest', 'quadriceps', 'hamstrings', 'lower-back')"
          },
          "category": {
            "type": "string",
            "description": "Muscle category for grouping (e.g. 'chest', 'back', 'shoulders', 'legs', 'arms', 'core', 'lower_back')"
          },
          "muscle_strength_score": {
            "type": "number",
            "description": "mscore: 0-100 effectiveness rating for building strength in target muscle. 100=most effective (e.g. Deadlift). Use this to prioritize exercise selection."
          },
          "sets_logged": {
            "type": "integer",
            "description": "Total sets logged across all Fitbod users \u2014 higher means more commonly used and validated"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the exercise, form cues, and benefits"
          },
          "source": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GymProfile": {
        "type": "object",
        "description": "A gym or training location with its available equipment. The primary gym profile is used to filter exercise recommendations.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Gym or location name (e.g. 'Home Gym', 'LA Fitness', 'Outdoor / Trail')"
          },
          "equipment": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of available equipment (e.g. ['Barbell', 'Dumbbells', 'Cable Machine', 'Pull Up Bar'])"
          },
          "is_primary": {
            "type": "boolean",
            "description": "If true, this is the default gym for exercise recommendations"
          },
          "notes": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GymProfileInput": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "equipment": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "is_primary": {
            "type": "boolean"
          },
          "notes": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ]
}